+++ react/typings/vtex.styleguide.d.ts
@@ -1,5 +1,6 @@
 declare module 'vtex.styleguide' {
   export const ButtonWithIcon
+  export const ButtonPlain
   export const IconClear
   export const IconPlusLines
 }